+++ bob/playbook/playbook.go
@@ -135,9 +135,6 @@ func (p *Playbook) TaskNeedsRebuild(taskname string, hashIn hash.In) (rebuildReq
 			return err
 		}
 
-		// TODO: In case the task does not exist check if a artifact can be used?
-		//       Part of no-permission-workflow.
-
 		// Ignore the task itself
 		if task.Name() == tn {
 			return nil
@@ -168,30 +165,32 @@ func (p *Playbook) TaskNeedsRebuild(taskname string, hashIn hash.In) (rebuildReq
 			// In case of a invalid traget a rebuild is required
 
 			// TODO: simplify verify by check size + modification time of target.
-			rebuildRequired = !target.Verify()
+			targetValid := target.VerifyShallow()
+			boblog.Log.V(3).Info(fmt.Sprintf("%-*s\t 11111 TargetValid is %t", p.namePad, coloredName, targetValid))
+			rebuildRequired = !targetValid
 
-			// Try to load a target from the store when a rebuild is required.
-			// If not assure the artifact exists in the store.
-			if rebuildRequired {
-				boblog.Log.V(2).Info(fmt.Sprintf("[task:%s] trying to get target from store", taskname))
-				ok, err := task.ArtifactUnpack(hashIn)
-				boblog.Log.Error(err, "Unable to get target from store")
+			// // Try to load a target from the store when a rebuild is required.
+			// // If not assure the artifact exists in the store.
+			// if rebuildRequired {
+			// 	boblog.Log.V(2).Info(fmt.Sprintf("[task:%s] trying to get target from store", taskname))
+			// 	ok, err := task.ArtifactUnpack(hashIn)
+			// 	boblog.Log.Error(err, "Unable to get target from store")
 
-				if ok {
-					rebuildRequired = false
-				} else {
-					boblog.Log.V(3).Info(fmt.Sprintf("[task:%s] failed to get target from store", taskname))
-				}
-			} else {
-				// Hint: Once there was a time when we created the target in the store
-				// in case no rebuild was required and the target doesn't exist.
-				// Though this should only be done after the target was really build..
-				// If loaded from a remote.. it anyway is synced through the local store.
-				if !task.ArtifactExists(hashIn) {
-					boblog.Log.V(3).Info(fmt.Sprintf("%-*s\tNEEDS REBUILD\t(target not in local store)", p.namePad, coloredName))
-					return true, TargetInvalid, err
-				}
-			}
+			// 	if ok {
+			// 		rebuildRequired = false
+			// 	} else {
+			// 		boblog.Log.V(3).Info(fmt.Sprintf("[task:%s] failed to get target from store", taskname))
+			// 	}
+			// } else {
+			// 	// Hint: Once there was a time when we created the target in the store
+			// 	// in case no rebuild was required and the target doesn't exist.
+			// 	// Though this should only be done after the target was really build..
+			// 	// If loaded from a remote.. it anyway is synced through the local store.
+			// 	if !task.ArtifactExists(hashIn) {
+			// 		boblog.Log.V(3).Info(fmt.Sprintf("%-*s\tNEEDS REBUILD\t(target not in local store)", p.namePad, coloredName))
+			// 		return true, TargetInvalid, err
+			// 	}
+			// }
 
 			if rebuildRequired {
 				boblog.Log.V(3).Info(fmt.Sprintf("%-*s\tNEEDS REBUILD\t(invalid targets)", p.namePad, coloredName))
@@ -200,100 +199,6 @@ func (p *Playbook) TaskNeedsRebuild(taskname string, hashIn hash.In) (rebuildReq
 	}
 
 	return rebuildRequired, TargetInvalid, err
-}
-
-func (p *Playbook) Play() (err error) {
-	return p.play()
-}
-
-func (p *Playbook) play() error {
-
-	if p.done {
-		return ErrDone
-	}
-
-	p.playMutex.Lock()
-	defer p.playMutex.Unlock()
-
-	if p.start.IsZero() {
-		p.start = time.Now()
-	}
-
-	// Walk the task chain and determine the next build task. Send it to the task channel.
-	// Returns `taskQueued` when a task has been send to the taskChannel.
-	// Returns `taskFailed` when a task has failed.
-	// Once it returns `nil` the playbook is done with it's work.
-	var taskQueued = fmt.Errorf("task queued")
-	var taskFailed = fmt.Errorf("task failed")
-	err := p.Tasks.walk(p.root, func(taskname string, task *Status, err error) error {
-		if err != nil {
-			return err
-		}
-
-		//boblog.Log.V(3).Info(fmt.Sprintf("%-*s\t walking", p.namePad, taskname))
-
-		switch task.State() {
-		case StatePending:
-			// Check if all dependent tasks are completed
-			for _, dependentTaskName := range task.Task.DependsOn {
-				t, ok := p.Tasks[dependentTaskName]
-				if !ok {
-					//fmt.Printf("Task %s does not exist", dependentTaskName)
-					return usererror.Wrap(boberror.ErrTaskDoesNotExistF(dependentTaskName))
-				}
-
-				state := t.State()
-				if state != StateCompleted && state != StateNoRebuildRequired {
-					// A dependent task is not completed.
-					// So this task is not yet ready to run.
-					return nil
-				}
-			}
-		case StateFailed:
-			return taskFailed
-		case StateCanceled:
-			return nil
-		case StateNoRebuildRequired:
-			return nil
-		case StateCompleted:
-			return nil
-		case StateRunning:
-			return nil
-		default:
-		}
-
-		// fmt.Printf("sending task %s to channel\n", task.Task.Name())
-		// setting the task start time before passing it to channel
-		task.Start = time.Now()
-		// TODO: for async assure to handle send to a closed channel.
-		_ = p.setTaskState(task.Name(), StateRunning, nil)
-		p.taskChannel <- task.Task
-		return taskQueued
-	})
-
-	// taskQueued => return nil (happy path)
-	// taskFailed => return PlaybookFailed (ErrFailed)
-	// default    => return err
-	if err != nil {
-		if errors.Is(err, taskQueued) {
-			return nil
-		}
-		if errors.Is(err, taskFailed) {
-			return ErrFailed
-		}
-		return err
-	}
-
-	// Avoid finishing the playbook before all task are done running
-	if p.numRunningTasks() > 0 {
-		return nil
-	}
-
-	// no work done, usually happens when
-	// no task needs a rebuild.
-	p.Done()
-
-	return nil
 }
 
 func (p *Playbook) numRunningTasks() int {
@@ -406,7 +311,7 @@ func (p *Playbook) TaskCompleted(taskname string, hashIn hash.In) (err error) {
 	if target != nil {
 		bi, err := target.BuildInfo()
 		errz.Fatal(err)
-		buildInfo.Target = bi
+		buildInfo.Target = *bi
 	}
 
 	// Store buildinfo
