+++ bobtask/artifact_info.go
@@ -14,15 +14,19 @@ type ArtifactInfo interface {
 // artifactInfo is a helper to debug artifacts
 // during testing or from the cli.
 type artifactInfo struct {
-	// targets contained
-	targets []string
+	// targetsFilesystem are stored locally on disk
+	targetsFilesystem []string
+
+	// targetsDocker stored in a local docker registry
+	targetsDocker []string
 
 	metadata *ArtifactMetadata
 }
 
 func newArtifactInfo() *artifactInfo {
 	ai := &artifactInfo{
-		targets: []string{},
+		targetsFilesystem: []string{},
+		targetsDocker:     []string{},
 	}
 	return ai
 }
@@ -37,9 +41,15 @@ func (ai *artifactInfo) String() string {
 	indent := "  "
 	fmt.Fprintf(buf, "%s\n", "Artifact Info")
 
-	fmt.Fprintf(buf, "%s%s\n", indent, "targets:")
+	fmt.Fprintf(buf, "%s%s\n", indent, "filesystem targets:")
 	i := indent + "  "
-	for _, t := range ai.targets {
+	for _, t := range ai.targetsFilesystem {
+		fmt.Fprintf(buf, "%s%s\n", i, t)
+	}
+
+	fmt.Fprintf(buf, "%s%s\n", indent, "docker targets:")
+	i = indent + "  "
+	for _, t := range ai.targetsDocker {
 		fmt.Fprintf(buf, "%s%s\n", i, t)
 	}
 
@@ -50,7 +60,7 @@ func (ai *artifactInfo) String() string {
 		fmt.Fprintf(buf, "%s%s%s\n", i, "inputHash: ", ai.metadata.InputHash)
 		fmt.Fprintf(buf, "%s%s%s\n", i, "project: ", ai.metadata.Project)
 		fmt.Fprintf(buf, "%s%s%s\n", i, "createdAt: ", ai.metadata.CreatedAt.Format(time.RFC822Z))
-		fmt.Fprintf(buf, "%s%s%s\n", i, "targetType: ", ai.metadata.TargetType)
+		//fmt.Fprintf(buf, "%s%s%s\n", i, "targetType: ", ai.metadata.TargetType)
 	}
 
 	return buf.String()
