+++ bobtask/artifact_metadata.go
@@ -2,8 +2,6 @@ package bobtask
 
 import (
 	"time"
-
-	"github.com/benchkram/bob/bobtask/targettype"
 )
 
 type ArtifactMetadata struct {
@@ -22,13 +20,13 @@ type ArtifactMetadata struct {
 
 	// TargetType sets the type of target, path/docker,
 	// default sets to `path`
-	TargetType targettype.T `yaml:"target_type,omitempty"`
+	//TargetType targettype.T `yaml:"target_type,omitempty"`
 }
 
 func NewArtifactMetadata() *ArtifactMetadata {
 	am := &ArtifactMetadata{
-		CreatedAt:  time.Now(),
-		TargetType: targettype.Path,
+		CreatedAt: time.Now(),
+		//	TargetType: targettype.Path,
 	}
 	return am
 }
