+++ bobtask/rebuild.go
@@ -30,8 +30,7 @@ func (t *Task) NeedsRebuild(options *RebuildOptions) (_ bool, err error) {
 		errz.Fatal(err)
 	}
 
-	_, err = t.buildInfoStore.GetBuildInfo(hashIn.String())
-	if err != nil {
+	if !t.buildInfoStore.BuildInfoExists(hashIn.String()) {
 		if errors.Is(err, buildinfostore.ErrBuildInfoDoesNotExist) {
 			boblog.Log.V(4).Info(fmt.Sprintf("%s, Searching for input hash %s failed", t.name, hashIn.String()))
 			return true, nil
@@ -39,20 +38,6 @@ func (t *Task) NeedsRebuild(options *RebuildOptions) (_ bool, err error) {
 		errz.Fatal(err)
 	}
 	boblog.Log.V(4).Info(fmt.Sprintf("%s, Searching for input hash %s succeeded", t.name, hashIn.String()))
-
-	// storedHashes, err := t.ReadHashes()
-	// if err != nil {
-	// 	if errors.Is(err, ErrHashesFileDoesNotExist) {
-	// 		return true, nil
-	// 	} else if errors.Is(err, ErrTaskHashDoesNotExist) {
-	// 		return true, nil
-	// 	} else {
-	// 		return true, fmt.Errorf("failed to read file hashes: %w", err)
-	// 	}
-	// }
-
-	// _, ok := storedHashes[*hashIn]
-	// return !ok, nil
 
 	return false, nil
 }
